Fix scheduler bug.
The scheduler’s first pass - assigning people to dates for tasks that are manually scheduled - needs to use the backwards calendar, not the forwards one. We don’t expect any conflicts until we do the forwards pass at the end (which uses a clean calendar rather than the one built up based on manually scheduled tasks and the backwards pass). The reason for this is it’s more useful to see conflicts between “start now” and “finish by X” rather than just having forward scheduled tasks slide to the very end of the project.