Better Ongoing / Rolling Project
Broadly, this change is aimed at support for ongoing / rolling projects.
Prior to this change, projects would quickly become problematic for the
optimizer, since it would roll the window back to accommodate all start
dates.

Now, we will run a pruning step before optimizing which:
1. Removes all tasks with end date before "today" from the optimization.
2. When start < today < end, adjust estimate to end - today and include
   in the optimization.

The normal highlighting Fantasia provides after the fact will still
work, but it forces the PM to keep the project up to date. The only
problematic case is where start and end date are before today but we
want it included in the optimization. In these cases, they'll be
highlighted red ( since the optimization will not force itself to step
backward to include them ), but the user is forced to set an end date
later than today to have it included, which they should since it's not
finished.